implements method update in SimpleGameObject, change speed in double and add testUpdate

+++ src/main/java/it/unibo/tankBattle/model/gameObject/api/component/Movable.java
@@ -8,7 +8,7 @@ public interface Movable extends Component {
      * 
      * @return the component speed
      */
-    public int getSpeed();
+    public double getSpeed();
 
     /**
      *
